Create 3-3-2
Write a program that takes input as an integer grade from the keyboard and translates it to a letter grade. For example, if a user enters 4, print "A", and if they enter 0 print "F."  You can assume that an integer between 0 and 4 will be input. To gain full credit for this exercise you will need to use else if statements when checking the grade.