Create 3-3-3
Create a program to let the user practice their multiplication tables. Print a multiplication question using two random integers between 1 and 12 as shown in the sample run. Then allow the user input an int as their answer to the multiplication of the two numbers. If they enter the correct product print “Correct!” otherwise print “Wrong”. You must use an else statement in your code to receive full credit for this exercise.