Create 3-3-4
We are going to revisit a former coding activity on temperature, with a slight adjustment. 

You are running an experiment that involves hatching chicken eggs. Bird eggs are very sensitive to temperature and chickens' eggs will hatch between 99 and 102 degrees Fahrenheit.

Write the code for the sensor that will be tracking the temperature. If the temperature falls below 99 or above 102 your code should print “WARNING” otherwise “Temperature is OK”. The values 99 and 102 are considered to be included in the ok range and the input will be a double.