Create 3-4-3
The Internet runs on web addresses. The addresses we type represent the IP address for each site and how the computer finds an individual web page.

IP addresses are made up of four numbers, each between 0 and 255 separated by a period. For example, 128.253.21.58 is an IP address.

Write a program which takes four integer inputs and tests if they make up a valid IP address. In other words, your program should test to see if each number entered is between 0 and 255 inclusive.

If any octet entered is incorrect your program should print "Octet # is incorrect" where # is the octet number (1 being the first,  4 the last). If all octets are correct, your program should print "IP Address: " followed by the correct IP address which is made up of the four octets separated by periods (.). See the sample runs for examples of this output.