Create 3-5-1
Debug the code provided in the starter file. This program is supposed to print "Ratio OK" if the ratio of two integers entered by the user is between 4 inclusive and 6 exclusive (i.e. if the second divided by the first is less than 6 and greater than or equal to 4). The debugged code should not produce any error messages as long as two integers are entered by the user.