Create 3-5-2
Debug the code in the starter file so it functions as intended and does not cause any errors. This program is intended to take two integer inputs and determine whether the second is a multiple of the first or not.

The number b is a multiple of the number a if b can be divided by a with no remainder. Remember though that no number can be divided by 0 - so no numbers should be considered a multiple of 0 for the purpose of this exercise.