Create Assignment 3
In this assignment, you will be creating a program that requires a secret code to “unlock.” The program should first welcome the user and ask the user to input his/her name. Then, the program will greet the user using the entered name and ask whether they are ready to crack the code. If they type "yes" (using any combination of upper/lower-case letters) then the program should proceed to the next stage. Otherwise, it should finish at that point and produce no more output.

In order to “crack the code,” the user must input three integer numbers which satisfy the following conditions:

The first number must be the number 3.
The second number can either be the number 1 or be between 33 and 100, inclusive.
The third number must be a positive number (remember 0 is not positive) that is either evenly divisible by 3 or evenly divisible by 7
As the user enters each number, the program should immediately check whether or not the number satisfies its respective requirement(s) listed above. If the requirement(s) is satisfied, then a “Correct!” confirmation should be printed out. If the requirement(s) is NOT satisfied, then an error message should be printed out, and the user cannot input any more numbers. (i.e. if the user input 5 as the first number, then the user has failed to crack the code, and cannot guess the second or third number.)

Please refer to the sample outputs below for formatting. (Any output statements from the program must match the sample formatting exactly.)